pipeline/requirements.txt
requests>=2.31.0
pandas>=2.0.0
supabase>=2.3.0
python-dotenv>=1.0.0
cbbpy>=1.2.0
pyarrow>=14.0.0
